% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_v1_x.R
\name{authenticate_v1_x}
\alias{authenticate_v1_x}
\title{Authenticate against the MySportsFeeds API (v1.x)}
\usage{
authenticate_v1_x(apikey, password)
}
\arguments{
\item{apikey.}{Your MySportsFeeds API Key}

\item{password.}{Your MySportsFeeds account password}
}
\description{
Sets the authentication credentials in R environment variables
}
\examples{
\dontrun{
authenticate_v1_x("<your-api-key>", "<your_account_password>")}
}
