% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_v2_x.R
\name{authenticate_v2_x}
\alias{authenticate_v2_x}
\title{Authenticate against the MySportsFeeds API (v2.x)}
\usage{
authenticate_v2_x(apikey)
}
\arguments{
\item{apikey.}{Your MySportsFeeds API Key}
}
\description{
Sets the authentication credentials in R environment variables
}
\examples{
\dontrun{
authenticate_v2_x("<your-api-key>")}
}
