% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate_v1_0}
\alias{authenticate_v1_0}
\title{Authenticate against the MySportsFeeds API (v1.0)}
\usage{
authenticate_v1_0(username, password)
}
\arguments{
\item{username.}{Your MySportsFeeds account username}

\item{password.}{Your MySportsFeeds account password}
}
\description{
Sets the authentication credentials in R environment variables
}
\examples{
\dontrun{
authenticate_v1_0("your-username", "your_password")}
}
