% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msf_get_results.R
\name{msf_get_results}
\alias{msf_get_results}
\title{Interface with the MySportsFeeds API}
\usage{
msf_get_results(version = "1.2", league, season, feed, params = list(),
  verbose = FALSE)
}
\arguments{
\item{version}{character. The version of the API. Defaults to "1.2".}

\item{league}{character. The league abbreviation. (e.g. - "nhl", "nfl", "nba", "mlb")}

\item{season}{character. The applicable season.  (e.g. - "2016-2017-regular")}

\item{feed}{character. The name of the requested feed.  (e.g. - "player_gamelogs")}

\item{params}{list. A list that defines additional parameters. Default is empty list.}

\item{verbose}{logical. For debugging, prints status messages to the console, which can be helpful for walking through results. Default is TRUE.}
}
\value{
a list.
}
\description{
Return data from the MySportsFeeds API
}
\examples{
\dontrun{
authenticate_v1_x("<your_api_key>", "<your_account_password>")
results <- msf_get_results(version = "1.2",
                           league = "nhl",
                           season = "2016-2017-regular",
                           feed = "player_gamelogs",
                           params = list()
                           verbose = TRUE)
}
\dontrun{
authenticate_v2_x("<your_api_key>")
results <- msf_get_results(version = "2.0",
                           league = "nhl",
                           season = "2016-2017-regular",
                           feed = "seasonal_player_gamelogs",
                           params = list()
                           verbose = TRUE)
}
}
